+++ libft/ft_memset.c
@@ -6,13 +6,15 @@
 /*   By: lduran-f <lduran-f@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/23 00:12:22 by lduran-f          #+#    #+#             */
-/*   Updated: 2025/09/23 00:41:01 by lduran-f         ###   ########.fr       */
+/*   Updated: 2025/10/14 18:22:39 by lduran-f         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
